use unique monitoring configs
